- Add AccountDiscussionLanguagesView, as it is a full view. - Move discussion languages logic to new view. - Add accountDiscussionLanguage route.
