+++ Mlem/Views/Tabs/Settings/Components/Views/Account/AccountGeneralSettingsView.swift
@@ -17,39 +17,15 @@ struct AccountGeneralSettingsView: View {
     @State var showBotAccounts: Bool = false
     @State var sendNotificationsToEmail: Bool = false
     
-    @State var discussionLanguages: Set<Int> = .init()
-    
     init() {
         if let info = siteInformation.myUserInfo {
             _showNsfw = State(wrappedValue: info.localUserView.localUser.showNsfw)
             _showBotAccounts = State(wrappedValue: info.localUserView.localUser.showBotAccounts)
             _sendNotificationsToEmail = State(wrappedValue: info.localUserView.localUser.sendNotificationsToEmail)
-            _discussionLanguages = State(wrappedValue: Set(info.discussionLanguages))
-        }
-        
-    }
-    
-    func saveDiscussionLanguages() {
-        let newValue = Array(discussionLanguages).sorted()
-        if newValue != siteInformation.myUserInfo?.discussionLanguages {
-            let oldValues = siteInformation.myUserInfo?.discussionLanguages ?? []
-            siteInformation.myUserInfo?.discussionLanguages = newValue
-            Task {
-                if let info = siteInformation.myUserInfo {
-                    do {
-                        try await apiClient.saveUserSettings(myUserInfo: info)
-                    } catch {
-                        discussionLanguages = Set(oldValues)
-                        siteInformation.myUserInfo?.discussionLanguages = oldValues
-                        errorHandler.handle(error)
-                    }
-                }
-            }
         }
     }
     
     var body: some View {
-
         Form {
             Section {
                 SwitchableSettingsItem(
@@ -100,46 +76,9 @@ struct AccountGeneralSettingsView: View {
                 }
             }
             Section {
-                NavigationLink("Discussion Languages") {
-                    Form {
-                        Section {
-                            Toggle(isOn: Binding(
-                                get: { discussionLanguages.contains(0) },
-                                set: { selected in
-                                    if selected {
-                                        discussionLanguages.insert(0)
-                                    } else {
-                                        discussionLanguages.remove(0)
-                                    }
-                                    saveDiscussionLanguages()
-                                }
-                            )) {
-                                Text("Undetermined")
-                            }
-                        } footer: {
-                            Text("If you deselect Undetermined, you won't see most content.")
-                        }
-                        Section {
-                            ForEach(siteInformation.allLanguages.dropFirst(), id: \.self) { language in
-                                Toggle(isOn: Binding(
-                                    get: { discussionLanguages.contains(language.id) },
-                                    set: { selected in
-                                        if selected {
-                                            discussionLanguages.insert(language.id)
-                                        } else {
-                                            discussionLanguages.remove(language.id)
-                                        }
-                                        saveDiscussionLanguages()
-                                    }
-                                )) {
-                                    Text(language.name)
-                                }
-                            }
-                        }
-                    }
-                    .fancyTabScrollCompatible()
+                NavigationLink(.settings(.accountDiscussionLanguages)) {
+                    Text("Discussion Languages")
                 }
-                
             }
             Section {
                 SwitchableSettingsItem(
